backend/requirements.txt
# FastAPI核心
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# 数据库
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.1

# Web3
web3==6.15.0
eth-account==0.11.0

# Redis
redis==5.0.1
aioredis==2.0.1

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 安全
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1

# 工具
python-dateutil==2.8.2
pytz==2024.1

# 日志
loguru==0.7.2

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # 用于测试API
